+++ src/SolrExpress/Search/Result/DateMath.cs
